move yaml function to shared
