+++ docker/lib/dependabot/docker_compose/file_fetcher.rb
@@ -8,9 +8,33 @@ module Dependabot
     class FileFetcher < Dependabot::Shared::SharedFileFetcher
       FILENAME_REGEX = /(docker-)?compose(?>\.[\w-]+)?\.ya?ml/i
 
+      sig { override.returns(T::Array[DependencyFile]) }
+      def fetch_files
+        fetched_files = correctly_encoded_docker_compose_files
+
+        return fetched_files if fetched_files.any?
+
+        raise_appropriate_error
+      end
+
       sig { override.returns(Regexp) }
       def self.filename_regex
         FILENAME_REGEX
+      end
+
+      def docker_compose_files
+        @docker_compose_files ||=
+          repo_contents(raise_errors: false).
+            select { |f| f.type == "file" && f.name.match?(FILENAME_REGEX) }.
+            map { |f| fetch_file_from_host(f.name) }
+      end
+
+      def correctly_encoded_docker_compose_files
+        docker_compose_files.select { |f| f.content.valid_encoding? }
+      end
+
+      def incorrectly_encoded_docker_compose_files
+        docker_compose_files.reject { |f| f.content.valid_encoding? }
       end
 
       sig { override.returns(String) }
