+++ docker/lib/dependabot/shared/utils/helpers.rb
@@ -4,7 +4,7 @@
 require "sorbet-runtime"
 
 module Dependabot
-  module Docker
+  module Shared
     module Utils
       HELM_REGEXP = /values[\-a-zA-Z_0-9]*\.ya?ml$/i
